requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
astroid==2.8.4
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==35.0.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.9
django-allauth==0.45.0
djangorestframework==3.12.4
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.3.2
future==0.18.2
httplib2==0.14.0
idna==3.3
importlib-metadata==1.5.0
isort==5.10.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.1
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.4.0
platformdirs==2.4.0
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.3.0
pylint==2.11.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2021.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
virtualenv==20.10.0
wadllib==1.3.3
wrapt==1.13.3
xkit==0.0.0
zipp==1.0.0
